Create midithru.py
This is a CircuitPython example that handles sending and receiving MIDI messages on a Pybadge with a MIDI FeatherWing attached. It uses the UART to receive and transmit the MIDI. The CircuitPython MIDI library was not used, because it apparently misses/ignores lots of the messages being received if MIDI Clock messages are being received (lots of messages per second.) This example filters out MIDI System messages as soon as they are received, and does not seem to drop any received NoteOn and CC messages.